+++ __tests__/test-error-messages.js
@@ -16,15 +16,19 @@
 
 // @flow
 
-import { ABIVersions, tests } from './_/init-tests';
+import { TONClientError } from '../src/TONClient';
+import { ABIVersions, nodeSe, tests } from './_/init-tests';
 import { TONMnemonicDictionary } from '../src/modules/TONCryptoModule';
 
 const WalletContractPackage = tests.loadPackage('WalletContract');
+const HelloContractPackage = tests.loadPackage('Hello');
+
+jest.setTimeout(100000);
 
 beforeAll(tests.init);
 afterAll(tests.done);
 
-async function expectError(code: number, source: string, message?: string, f) {
+async function expectError(code: number, source: ?string, message: ?string, f: () => Promise<void>) {
     try {
         await f();
         //$FlowFixMe
@@ -32,11 +36,11 @@ async function expectError(code: number, source: string, message?: string, f) {
     } catch (error) {
         expect({
             code: error.code,
-            source: error.source
+            ...(source ? { source: error.source } : {})
         })
             .toEqual({
                 code,
-                source
+                ...(source ? { source } : {})
             });
         if (message) {
             expect(error.message)
@@ -44,6 +48,68 @@ async function expectError(code: number, source: string, message?: string, f) {
         }
     }
 }
+
+async function expectErrorCode(code: number, f: () => Promise<void>) {
+    return expectError(code, null, null, f);
+}
+
+test.each(ABIVersions)('Detailed errors (ABI v%i)', async (abiVersion) => {
+    const { contracts, crypto } = tests.client;
+    tests.client.config.data.waitForTimeout = 5000;
+    const helloPackage = HelloContractPackage[abiVersion];
+
+    let helloKeys = await crypto.ed25519Keypair();
+    let helloAddress = (await contracts.createDeployMessage({
+        package: helloPackage,
+        constructorParams: {},
+        keyPair: helloKeys,
+    })).address;
+
+    await expectErrorCode(TONClientError.code.ACCOUNT_MISSING, async () => {
+        await contracts.deploy({
+            package: helloPackage,
+            constructorParams: {},
+            keyPair: helloKeys,
+        });
+    });
+
+    await tests.get_grams_from_giver(helloAddress, 100);
+    await expectErrorCode(TONClientError.code.ACCOUNT_BALANCE_TOO_LOW, async () => {
+        await contracts.deploy({
+            package: helloPackage,
+            constructorParams: {},
+            keyPair: helloKeys,
+        });
+    });
+
+    helloKeys = await crypto.ed25519Keypair();
+    helloAddress = (await contracts.createDeployMessage({
+        package: helloPackage,
+        constructorParams: {},
+        keyPair: helloKeys,
+    })).address;
+
+    await expectErrorCode(TONClientError.code.ACCOUNT_MISSING, async () => {
+        await contracts.run({
+            address: helloAddress,
+            abi: helloPackage.abi,
+            functionName: 'touch',
+            input: {},
+            keyPair: helloKeys,
+        });
+    });
+
+    await tests.get_grams_from_giver(helloAddress, 100);
+    await expectErrorCode(TONClientError.code.ACCOUNT_CODE_MISSING, async () => {
+        await contracts.run({
+            address: helloAddress,
+            abi: helloPackage.abi,
+            functionName: 'touch',
+            input: {},
+            keyPair: helloKeys,
+        });
+    });
+});
 
 test.each(ABIVersions)('Test SDK Errors 1-3 (ABI v%i)', async (abiVersion) => {
     const { contracts, crypto } = tests.client;
@@ -128,6 +194,35 @@ test.each(ABIVersions)('Test SDK Errors 1-3 (ABI v%i)', async (abiVersion) => {
             functionName: 'participant_list',
         });
     });
+});
+const literallyJustDateNow = () => Date.now();
+
+test.each(ABIVersions)('Test SDK Error 1013', async (abiVersion) => {
+    if (!nodeSe) {
+        const { crypto } = tests.client;
+        const helloKeys = await crypto.ed25519Keypair();
+        const helloPackage = HelloContractPackage[abiVersion];
+
+        const realDateNow = Date.now.bind(global.Date);
+        const start = Date.now() - 20000;
+        const dateNowStub = jest.fn(() => start);
+        global.Date.now = dateNowStub;
+
+        expect(literallyJustDateNow())
+            .toBe(start);
+        expect(dateNowStub)
+            .toHaveBeenCalled();
+
+        await expectError(1013, 'client',
+            'You local clock is out of sync with the server time. It is a critical condition for sending messages to the blockchain. Please sync you clock with the internet time', async () => {
+                await tests.deploy_with_giver({
+                    package: helloPackage,
+                    constructorParams: {},
+                    keyPair: helloKeys,
+                });
+            });
+        global.Date.now = realDateNow;
+    }
 });
 
 test.each(ABIVersions)('Test SDK Errors > 2000 (ABI v%i)', async (abiVersion) => {
@@ -236,30 +331,34 @@ test.each(ABIVersions)('Test SDK Errors > 2000 (ABI v%i)', async (abiVersion) =>
             .toMatch('Invalid bip39 entropy:');
     }
 
-    try {
-        await crypto.hdkeyXPrvDerivePath('???', '', true);
-    } catch (error) {
-        expect(error.source)
-            .toEqual('client');
-        expect(error.code)
-            .toEqual(2018);
-    }
-    try {
+    await expectError(2017, 'client', 'Invalid bip39 phrase: one two', async () => {
+        await crypto.mnemonicDeriveSignKeys({
+            phrase: 'one two',
+        });
+    });
+
+    await expectError(2017, 'client', 'Invalid bip39 phrase:', async () => {
+        await crypto.mnemonicDeriveSignKeys({
+            phrase: '',
+        });
+    });
+
+    await expectError(2018, 'client', 'Invalid bip32 key: ', async () => {
+        await crypto.hdkeyXPrvDerivePath('', '', true);
+    });
+
+    await expectError(2019, 'client', 'Invalid bip32 derive path:', async () => {
+        await crypto.hdkeyXPrvDerivePath('xprv9s21ZrQH143K25JhKqEwvJW7QAiVvkmi4WRenBZanA6kxHKtKAQQKwZG65kCyW5jWJ8NY9e3GkRoistUjjcpHNsGBUv94istDPXvqGNuWpC',
+            'm/44\'/6   0\'/0\'/0\'', false);
+    });
+
+    await expectError(2022, 'client', 'Invalid mnemonic dictionary', async () => {
         await crypto.mnemonicFromRandom({
             // $FlowFixMe
             dictionary: 255,
             wordCount: 12,
         });
-    } catch (error) {
-        expect(error.source)
-            .toEqual('client');
-        expect(error.code)
-            .toEqual(2022);
-        expect(error.data)
-            .toBeNull();
-        expect(error.message)
-            .toMatch('Invalid mnemonic dictionary');
-    }
+    });
 
     for (const dict in TONMnemonicDictionary) {
         try {
@@ -294,6 +393,7 @@ test.each(ABIVersions)('Test SDK Errors 3000-3020 (ABI v%i)', async (abiVersion)
     await expectError(3002, 'client', 'Invalid contract image:', async () => {
         await contracts.createDeployMessage({
             package: {
+                //$FlowFixMe
                 abi: '',
                 imageBase64: '#',
             },
@@ -304,6 +404,7 @@ test.each(ABIVersions)('Test SDK Errors 3000-3020 (ABI v%i)', async (abiVersion)
     await expectError(3003, 'client', 'Image creation failed: failed to fill whole buffer', async () => {
         await contracts.createDeployMessage({
             package: {
+                //$FlowFixMe
                 abi: '',
                 imageBase64: '',
             },