+++ babel.config.js
@@ -22,14 +22,10 @@ const plugins = [
     ['@babel/plugin-transform-async-to-generator'],
     ['@babel/plugin-transform-block-scoping'],
     ['@babel/plugin-transform-classes'],
+    ['@babel/plugin-transform-runtime', { helpers: false }]
 ];
 
 module.exports = {
     presets,
     plugins,
-    env: {
-        test: {
-            plugins: ["@babel/plugin-transform-runtime"]
-        }
-    }
 };